+++ bin/pack.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-# Weekday Plasmoid for KDE
+# Weekday Widget for KDE
 #
 # @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
 # @copyright 2020 Marcin Orlowski
@@ -17,9 +17,10 @@ declare -r ROOT_DIR="$(realpath "$(dirname "$(realpath "${0}")")/..")"
 source "${ROOT_DIR}/bin/common.sh"
 
 function packPlasmoid() {
+	local -r DIR="$(pwd)"
 	pushd "${ROOT_DIR}/src" > /dev/null
 
-	local -r plasmoid_file="$(pwd)/$(getPlasmoidFileName)"
+	local -r plasmoid_file="${DIR}/$(getPlasmoidFileName)"
 	zip -q -r "${plasmoid_file}" -- *
 	ls -ld "${plasmoid_file}"