+++ src/contents/ui/Day.qml
@@ -1,25 +1,21 @@
 /**
- * Weekday Widget
- *
- * Configurabler vertical multi clock plasmoid.
+ * Weekday Widget for KDE
  *
  * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
  * @copyright 2020 Marcin Orlowski
  * @license   http://www.opensource.org/licenses/mit-license.php MIT
- * @link      https://github.com/MarcinOrlowski/html-clock-plasmoid
+ * @link      https://github.com/MarcinOrlowski/weekday-plasmoid
  */
 
 import QtQuick 2.1
 import QtQuick.Layouts 1.1
-//import org.kde.plasma.components 3.0 as PlasmaComponents
-
 
 ColumnLayout {
-	property int dayIndex: 0
 	property string label: '?'
-	property bool highlight: false
-	property string highlightFg: "white"
-	property string highlightBg: "black"
+	property string fg: "FFffffff"
+	property string bg: "#00000000"
+	property bool bold: false
+	property bool italic: false
 
 	Layout.fillWidth: true
 	Layout.minimumWidth: weekGrid.cellMinWidth
@@ -29,17 +25,20 @@ ColumnLayout {
 	
 	Rectangle {
 		anchors.fill: parent
-		color: highlight ? highlightBg : "#00000000"
+		color: bg
 	}
 
 	Text {
-//		padding: 6
-		readonly property int dayIndex: 0
+		text: label
+		color: fg
+		font.bold: bold
+		font.italic: italic
 
-		Layout.alignment: Qt.AlignVCenter | Qt.AlignHCenter
-		text: parent.label
-		color: highlight ? highlightFg : "white"
+		Layout.fillWidth: true
+		Layout.margins: 4
+
 		horizontalAlignment: Text.AlignHCenter
 		verticalAlignment: Text.AlignVCenter
+		Layout.alignment: Qt.AlignVCenter | Qt.AlignHCenter
 	}
 }