+++ CHANGES.md
@@ -1,3 +1,10 @@
+* v1.2.0 (2020-12-17)
+  * Added widget tooltip with current date
+  * Added option to override used locale
+  * Added customizable week start day
+  * Appearance can now be configured
+  * Separate look config for past, current and future days
+
 * v1.1.0 (2020-12-16)
   * General code cleamup
   * Week layout not uses GridLayout for better positioning